fix: add null check into services

+++ src/service/epicService/AbstractEpicService.java
@@ -13,6 +13,7 @@ import util.TaskManagerConfig;
 import util.TaskStatus;
 
 import java.util.ArrayList;
+import java.util.NoSuchElementException;
 
 public abstract class AbstractEpicService implements EpicService {
     private final EpicRepository epicRepo;
@@ -29,6 +30,10 @@ public abstract class AbstractEpicService implements EpicService {
 
     @Override
     public Epic create(EpicCreationData data) {
+        if (data == null) {
+            data = new EpicCreationData(null, null);
+        }
+
         Integer id = idGenerator.generateNewId();
         Epic epic = new Epic(id, data);
         epic.setStatus(TaskStatus.NEW);
@@ -39,6 +44,11 @@ public abstract class AbstractEpicService implements EpicService {
     @Override
     public Epic get(Integer id) {
         Epic epic = epicRepo.get(id);
+
+        if (epic == null) {
+            throw new NoSuchElementException("epic not found");
+        }
+
         historyManager.add(epic);
         return epic;
     }
@@ -50,7 +60,15 @@ public abstract class AbstractEpicService implements EpicService {
 
     @Override
     public Epic update(EpicUpdationData data) {
+        if (data == null) {
+            throw new IllegalArgumentException();
+        }
+
         Epic savedEpic = epicRepo.get(data.getId());
+
+        if (savedEpic == null) {
+            throw new NoSuchElementException("epic not found");
+        }
 
         savedEpic.setName(data.getName());
         savedEpic.setDescription(data.getDescription());
@@ -61,6 +79,10 @@ public abstract class AbstractEpicService implements EpicService {
     @Override
     public void remove(Integer id) {
         Epic epic = epicRepo.get(id);
+
+        if (epic == null) {
+            return;
+        }
 
         for (Subtask subtask : epic.getSubtasks()) {
             subtaskRepo.remove(subtask.getId());
@@ -78,6 +100,11 @@ public abstract class AbstractEpicService implements EpicService {
     @Override
     public ArrayList<Subtask> getSubtasks(Integer epicId) {
         Epic epic = epicRepo.get(epicId);
+
+        if (epic == null) {
+            throw new NoSuchElementException("epic not found");
+        }
+
         return epic.getSubtasks();
     }
 }
